installed syntax highlighting library
